requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.7
django-admin-tools==0.9.0
django-celery-beat==2.0.0
django-crispy-forms==1.9.1
django-floppyforms==1.9.0
django-fluent-dashboard==1.0.1
django-hosts==4.0
django-js-asset==1.2.2
django-markdown2==0.3.1
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-timezone-field==4.0
django-widget-tweaks==1.4.8
future==0.18.2
html5lib==1.0.1
importlib-metadata==1.6.0
jsonfield==3.1.0
kombu==4.6.8
markdown2==2.4.0
Pillow==8.2.0
pycparser==2.20
Pyphen==0.9.5
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.2
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
tinycss2==1.0.2
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
zipp==3.1.0
